# api/openapi.yaml
openapi: "3.0.3"
info:
  title: "Debuglet dispatcher HTTP API"
  version: "1.2"
  description: "Wire contract of the dispatcher's public HTTP API. Bandwidth values are bits per second, timeouts are milliseconds, TESLA delays are seconds, TESLA anchor timestamps are Unix nanoseconds and every other timestamp is Unix seconds unless a field says otherwise. A client states the contract version it requires with the Debuglet-API-Version request header, declared on every operation below; every response carries the version this dispatcher implements in the same header. Requests are authenticated with a server-issued session: a native client presents it as an Authorization bearer token, a browser client in the session_token cookie plus the X-Debuglet-CSRF header on state-changing requests. Each operation states its own 401 and 403. Access to another account's run or payment order answers 404, so a response is no existence oracle. A dispatcher explicitly configured for the local development profile additionally serves the unauthenticated routes below as its own local operator. See docs/API.md for the access matrix and the compatibility and deprecation policy."
servers:
  - url: "/"
    description: "Dispatcher served at the root of its origin."
  - url: "/api"
    description: "Dispatcher served behind an explicit /api path prefix. Every route below is relative to the selected server URL."
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /debuglet:
    put:
      operationId: "submitDebuglets"
      summary: "Submit a paid batch of debuglets."
      description: "Admits the debuglets of one completed transaction. The server hashes the decoded debuglets array and rejects any batch that differs from the one the payment intent was created for. The encoded request body must not exceed 33554432 bytes."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitDebugletsRequest"
      parameters:
        - name: "Debuglet-API-Version"
          in: "header"
          required: false
          description: "Contract version the client requires, as major or major.minor. A version this dispatcher does not serve is answered 400 with code unsupported_api_version; an absent or empty value states no requirement. The discovery routes /version and /openapi.yaml and the health routes answer whatever this header says."
          schema:
            type: string
      responses:
        "200":
          description: "The debuglet IDs, in the order of the submitted batch."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugletIDList"
        "400":
          description: "Malformed body (invalid_request), a batch that does not match the intent (intent_mismatch), an unpaid transaction (payment_incomplete), an unsupported payment method (unsupported_payment_method), an invalid debuglet (invalid_request) or a policy the dispatcher does not admit, including a reserved window that does not fit (invalid_policy)."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Unknown transaction, wrong auth key, a transaction that belongs to another account, or a missing, expired or revoked credential (unauthorized). The four are deliberately indistinguishable."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "The scheduler cannot admit the batch within the executor's remaining capacity (capacity_exhausted)."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "The batch was rejected by the dispatcher after admission (internal_error)."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: "The transaction is backed by a chain payment method while blockchain payments are disabled (payments_disabled), or an operator has stopped the admission of new submissions on this dispatcher (service_unavailable). Nothing is admitted or scheduled either way; a payment order that was already paid is refunded, which spends it, so the same batch is refused from then on. When the refund could not be performed the message says the order is still paid and the batch can be submitted again once admission resumes."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: "cancelDebuglet"
      summary: "Ask the dispatcher to abort a debuglet."
      description: "A 204 acknowledges the cancellation request. It does not certify remote termination and does not rewrite an already recorded terminal result."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebugletDeleteRequest"
      parameters:
        - name: "Debuglet-API-Version"
          in: "header"
          required: false
          description: "Contract version the client requires, as major or major.minor. A version this dispatcher does not serve is answered 400 with code unsupported_api_version; an absent or empty value states no requirement. The discovery routes /version and /openapi.yaml and the health routes answer whatever this header says."
          schema:
            type: string
      responses:
        "204":
          description: "Cancellation acknowledged."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
        "400":
          description: "Malformed body (invalid_request), or the dispatcher refused the cancellation (cancel_refused)."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "The request presented no credential, or one that is unknown, expired or revoked (unauthorized). A dispatcher serving the local development profile also serves this route without a credential."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "No debuglet with that ID is recorded, or it belongs to another account. The two are deliberately indistinguishable; neither is cancelled."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /debuglet/{id}/state:
    get:
      operationId: "getDebugletState"
      summary: "Report the recorded state of one debuglet."
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Canonical 36-character debuglet UUID."
          schema:
            type: string
            format: uuid
        - name: "Debuglet-API-Version"
          in: "header"
          required: false
          description: "Contract version the client requires, as major or major.minor. A version this dispatcher does not serve is answered 400 with code unsupported_api_version; an absent or empty value states no requirement. The discovery routes /version and /openapi.yaml and the health routes answer whatever this header says."
          schema:
            type: string
      responses:
        "200":
          description: "The recorded state."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugletStateResponse"
        "400":
          description: "The path segment is not a canonical UUID."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "The request presented no credential, or one that is unknown, expired or revoked (unauthorized). A dispatcher serving the local development profile also serves this route without a credential."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "No debuglet with that ID is recorded, or it belongs to another account. The two are deliberately indistinguishable."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "The state could not be read."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /debuglet/{id}/logs:
    get:
      operationId: "getDebugletLogs"
      summary: "Read one page of a debuglet's stored output."
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Canonical 36-character debuglet UUID."
          schema:
            type: string
            format: uuid
        - name: "after"
          in: "query"
          required: false
          description: "Cursor: return entries with an ID strictly above this value. Defaults to 0. A present but malformed or negative value is rejected."
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: "limit"
          in: "query"
          required: false
          description: "Maximum number of entries. Defaults to 100. A present but malformed or non-positive value is rejected; a larger value is clamped to 1000."
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 1000
        - name: "Debuglet-API-Version"
          in: "header"
          required: false
          description: "Contract version the client requires, as major or major.minor. A version this dispatcher does not serve is answered 400 with code unsupported_api_version; an absent or empty value states no requirement. The discovery routes /version and /openapi.yaml and the health routes answer whatever this header says."
          schema:
            type: string
      responses:
        "200":
          description: "One page of stored output."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugletLogsResponse"
        "400":
          description: "The path segment is not a canonical UUID, or a pagination parameter is invalid."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "The request presented no credential, or one that is unknown, expired or revoked (unauthorized). A dispatcher serving the local development profile also serves this route without a credential."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "No debuglet with that ID is recorded, or it belongs to another account. The two are deliberately indistinguishable."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "The logs could not be read."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /list-debuglets:
    get:
      operationId: "listUserDebuglets"
      summary: "List the debuglets of the calling user."
      description: "Requires the session_token cookie of an existing user."
      parameters:
        - name: "limit"
          in: "query"
          required: false
          description: "Maximum number of rows. Defaults to 100; a larger value is clamped to 100."
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: "offset"
          in: "query"
          required: false
          description: "Number of rows to skip. Defaults to 0."
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: "Debuglet-API-Version"
          in: "header"
          required: false
          description: "Contract version the client requires, as major or major.minor. A version this dispatcher does not serve is answered 400 with code unsupported_api_version; an absent or empty value states no requirement. The discovery routes /version and /openapi.yaml and the health routes answer whatever this header says."
          schema:
            type: string
      responses:
        "200":
          description: "The user's debuglets, newest first."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugletList"
        "400":
          description: "An invalid limit or offset."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "The request did not authenticate as an account (unauthorized). This route reports one account's own runs, so the local development profile does not serve it without a credential either."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "The rows could not be read."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      description: "A session token from /auth/login, presented as \"Authorization: Bearer <token>\". This is what a native client uses; it is not ambient authority, so it needs no CSRF token."
    cookieAuth:
      type: "apiKey"
      in: "cookie"
      name: "session_token"
      description: "The same session token in a cookie, for a browser client. A state-changing request authenticated this way must repeat the session's CSRF token, returned as csrf_token and set in the session_csrf cookie, in the X-Debuglet-CSRF header."
  schemas:
    Error:
      type: object
      description: "Error envelope of every failure. Branch on code; message is a bounded diagnostic for a human and must not be parsed. A minor version may add codes, so treat an unknown code as a plain failure of the status it arrived with."
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: "Stable machine-readable failure code."
          enum:
            - "invalid_request"
            - "invalid_policy"
            - "unknown_executor"
            - "intent_mismatch"
            - "payment_incomplete"
            - "unsupported_payment_method"
            - "payments_disabled"
            - "unsupported_api_version"
            - "unauthorized"
            - "forbidden"
            - "not_found"
            - "capacity_exhausted"
            - "cancel_refused"
            - "method_not_allowed"
            - "unsupported_media_type"
            - "payload_too_large"
            - "internal_error"
            - "service_unavailable"
        message:
          type: string
          description: "Bounded human-readable diagnostic. Never carries a credential, a stack trace or internal database or transport detail."
...
    SubmitDebugletsRequest:
      type: object
      additionalProperties: false
      required:
        - debuglets
        - transaction_id
        - auth_key
      properties:
        debuglets:
          type: array
          description: "The identical debuglets array the payment intent was created for."
          items:
            $ref: "#/components/schemas/DebugletRequest"
        transaction_id:
          type: string
        auth_key:
          type: string
          description: "Auth key of the transaction. Empty for TEST transactions."
    DebugletIDList:
      type: array
      description: "Debuglet IDs in the order of the submitted batch."
      items:
        type: string
        format: uuid
    DebugletDeleteRequest:
      type: object
      additionalProperties: false
      required:
        - debuglet_id
        - executor_id
      properties:
        debuglet_id:
          type: string
          format: uuid
        executor_id:
          type: string
    DebugletStateResponse:
      type: object
      additionalProperties: false
      required:
        - state
        - error
        - executor_id
      properties:
        state:
          type: string
          description: "Opaque run state. RunStateExited is terminal."
        error:
          type: string
          description: "Recorded workload error. Empty when the run has not failed."
        executor_id:
          type: string
    DebugletLogsResponse:
      type: object
      additionalProperties: false
      required:
        - state
        - after
        - logs
        - has_more
      properties:
        state:
          type: string
          description: "Opaque run state at the time of the query."
        error:
          type: string
          description: "Recorded workload error. Absent when empty."
        after:
          type: integer
          format: int64
          description: "Cursor for the next page: the ID of the last returned entry, or the requested cursor for an empty page."
        logs:
          type: array
          nullable: true
          description: "Stored output entries with strictly increasing IDs. Null when the page is empty."
          items:
            $ref: "#/components/schemas/DebugletLogEntry"
        has_more:
          type: boolean
          description: "True when the page was filled to the limit. A full page may report true although no later entry exists."
    DebugletLogEntry:
      type: object
      additionalProperties: false
      required:
        - id
        - timestamp
        - output
      properties:
        id:
          type: integer
          format: int64
        timestamp:
          type: string
          description: "Server timestamp in UTC, formatted as 2006-01-02T15:04:05Z."
        output:
          type: string
          format: byte
          description: "Base64-encoded guest output bytes."
...
    DebugletList:
      type: array
      items:
        $ref: "#/components/schemas/DebugletResponse"
...